% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agg_fun1.R
\name{Agg_fun1}
\alias{Agg_fun1}
\title{Agg_fun1}
\usage{
Agg_fun1(B, x.val, y.val, const = 2)
}
\arguments{
\item{B}{A matrix of target-only estimator and target estimator from each source in each column}

\item{x.val, y.val}{validation X and y}

\item{const}{A parameter, by default is 2}
}
\value{
Aggregated beta estimator
}
\description{
Aggregating beta estimator
}
