% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_curve.R
\name{area_under_curve}
\alias{area_under_curve}
\title{area_under_curve}
\usage{
area_under_curve(y, y.prob)
}
\arguments{
\item{y, y.prob}{estimated y and actual y}
}
\value{
AUC
}
\description{
Calculating AUC
}
