% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commute.R
\name{commute}
\alias{commute}
\title{commute}
\usage{
commute(X, y, X.src, y.src, n.src, w.src, r, S)
}
\arguments{
\item{X}{Variables from the target. The variables need to be completely the same set and in the same order as variables used in the model parameter estimators.}

\item{y}{Response from the target.}

\item{X.src}{A list of shareable sources X}

\item{y.src}{A list of shareable sources y}

\item{n.src}{Sample size of K sources}

\item{w.src}{A list of source estimators from which individual level data cannot be shared}

\item{r}{A positive integer}

\item{S}{The first S commute estimators to be incorporated in the aggregation}
}
\value{
A list of two beta estimator using inverse weight and logistic regression respectively
}
\description{
Derive calibration term estimator and an estimator for the target parameter in each source sites
}
