% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_ridge_1D.R
\name{CV_ridge_1D}
\alias{CV_ridge_1D}
\title{CV_ridge_1D}
\usage{
CV_ridge_1D(x_train, y_train, w, rho, var)
}
\arguments{
\item{x_train}{x for training}

\item{y_train}{y for training}

\item{w}{Source model parameter estimators}

\item{rho}{Correlation of target and source model parameter estimators}

\item{var}{Variance of target and source model parameter estimators}
}
\value{
A list with the best MSE, beta and lambda
}
\description{
3-fold cross validation for selecting tunning parameters
}
