% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transRF.R
\docType{data}
\name{data_transRF}
\alias{data_transRF}
\title{Simulated data for transRF}
\format{
A list with target x and y, pre-trained random forest model and feature importance from source
}
\description{
Simulated data for transRF
}
