% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTL.R
\name{distTL}
\alias{distTL}
\title{distTL}
\usage{
distTL(X, y, w.src)
}
\arguments{
\item{X}{Variables from the target. The variables need to be completely the same set and in the same order as variables used in the model parameter estimators.}

\item{y}{Response from the target.}

\item{w.src}{Pre-trained source model parameter estimators.}
}
\value{
A list of effect estimator and tuning parameters lambda and eta from distTL.
}
\description{
Distanced-based transfer learning
}
