% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_squared_error.R
\name{mean_squared_error}
\alias{mean_squared_error}
\title{mean_squared_error}
\usage{
mean_squared_error(y_est, y_test)
}
\arguments{
\item{y_est, y_test}{estimated y and actual y}
}
\value{
MSE
}
\description{
Calculating MSE
}
