% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_closed_form_result.R
\name{ridge_closed_form_result}
\alias{ridge_closed_form_result}
\title{ridge_closed_form_result}
\usage{
ridge_closed_form_result(x, y, lam, w, eta, x_val, y_val, standard)
}
\arguments{
\item{x}{x for training}

\item{y}{y for training}

\item{lam, eta}{Tuning parameter}

\item{w}{Source model parameter estimates}

\item{x_val}{x for validation}

\item{y_val}{y for validation}

\item{standard}{TRUE: run traditional ridge regression}
}
\value{
A list with the beta and MSE
}
\description{
Calculating closed form result in a ridge regression
}
