% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_init.R
\name{ST_init}
\alias{ST_init}
\title{ST_init}
\usage{
ST_init(X.tar, y.tar)
}
\arguments{
\item{X.tar}{Target X}

\item{y.tar}{Target y}
}
\value{
A list of beta and tuning parameter lambda
}
\description{
Derive beta estimator from target data
}
