% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thres.R
\name{thres}
\alias{thres}
\title{thres}
\usage{
thres(b, k, p)
}
\arguments{
\item{b}{Original calibration estimator}

\item{k}{Target sample size}

\item{p}{Dimension}
}
\value{
Calibration estimator after adding threshold
}
\description{
Shrink estimators of somr sites with performance lower than a certain threshold to zero
}
