% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_init.R
\name{TL_init}
\alias{TL_init}
\title{TL_init}
\usage{
TL_init(X.tar, y.tar, X.src = NULL, y.src = NULL, w = NULL)
}
\arguments{
\item{X.tar}{Target X}

\item{y.tar}{Target y}

\item{X.src}{Source X if available, by default is NULL}

\item{y.src}{Source y if available, by default is NULL}

\item{w}{Source estimator if provided, by default is NULL}
}
\value{
A list of source estimator, calibration term estimator and an estimator for the target parameter
}
\description{
Derive calibration term estimator and an estimator for the target parameter in each source sites
}
