% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{train_test_split}
\usage{
train_test_split(x, y, size)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{size}{The ratio to split for training data}
}
\value{
A list of x_train, y_train, x_test, y_test
}
\description{
Spliting the training dataset and the testing dataset
}
