% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trans_global.R
\name{Trans_global}
\alias{Trans_global}
\title{Trans_global}
\usage{
Trans_global(X.tar, y.tar, X.src, y.src, delta = NULL)
}
\arguments{
\item{X.tar}{Variables from the target. The variables need to be completely the same set and in the same order as variables used in the model parameter estimators.}

\item{y.tar}{Response from the target.}

\item{X.src}{A list of shareable source X, the length of the list should be M}

\item{y.src}{A list of shareable source y, the length of the list should be M}

\item{delta}{A vector of calibration estimator for each source}
}
\value{
A list of source estimator, calibration term estimator and an estimator for the target parameter
}
\description{
Derive calibration term estimator and an estimator for the target parameter in each source sites
}
