% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transRF.R
\name{transRF}
\alias{transRF}
\title{transRF}
\usage{
transRF(X, y, X.test = NULL, rf.src, S, p.val = 0.1)
}
\arguments{
\item{X}{Variables from the target. The variables need to be completely the same set and in the same order as variables used in the source model.}

\item{y}{Response from the target.}

\item{X.test}{If no testing data is availble, automatically split 20\% target data for testing.}

\item{rf.src}{Trained random forest model from the source.}

\item{S}{Feature importance score from the source model.}

\item{p.val}{The percent of spliting training data for validation. The default value is 10\%.}
}
\value{
A list of predicted y and y.test if X.test is not provided.
}
\description{
a transfer learning random forest framework
}
